pyproject.toml
[tool.poetry]
name = "meow"
version = "0.1.0"
description = ""
authors = ["MysteryCoder456 <rehatbir.singh@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pygame-ce = "^2.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
